[REFACTOR] global 패키지로 분리
