+++ src/main/java/kr/sesac/aoao/server/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package kr.sesac.aoao.server.config;
+package kr.sesac.aoao.server.global.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
