+++ src/main/java/kr/sesac/aoao/server/global/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package kr.sesac.aoao.server.exception;
+package kr.sesac.aoao.server.global.exception;
 
 import org.springframework.http.HttpStatus;
