+++ src/main/java/kr/sesac/aoao/server/global/exception/GlobalExceptionHandler.java
@@ -1,11 +1,11 @@
-package kr.sesac.aoao.server.exception;
+package kr.sesac.aoao.server.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import kr.sesac.aoao.server.controller.dto.response.ExceptionResponse;
+import kr.sesac.aoao.server.global.controller.dto.response.ExceptionResponse;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
